+++ samplesheets/views.py
@@ -37,7 +37,7 @@ from .models import (
     GenericMaterial,
 )
 from .rendering import SampleSheetTableBuilder, EMPTY_VALUE
-from .tasks import update_project_cache
+from .tasks import update_project_cache_task
 from .utils import get_sample_dirs, compare_inv_replace, get_sheets_url
 
 
@@ -478,24 +478,10 @@ class SampleSheetCacheUpdateView(
 
     def get(self, request, *args, **kwargs):
         project = self.get_project()
-        timeline = get_backend_api('timeline_backend')
-        tl_event = None
 
-        if timeline:
-            tl_event = timeline.add_event(
-                project=project,
-                app_name=APP_NAME,
-                user=request.user,
-                event_name='sheet_cache_update',
-                description='update cache for project sheets',
-                status_type='SUBMIT',
-                status_desc='Asynchronous update initiated',
-            )
-
-        update_project_cache.delay(
+        update_project_cache_task.delay(
             project_uuid=str(project.sodar_uuid),
             user_uuid=str(request.user.sodar_uuid),
-            tl_uuid=str(tl_event.sodar_uuid),
         )
 
         messages.warning(